Moved authentication services into a newer namespace and separate folder for clarity.
